Move utilities to separate file
